Presence detection implementation (#8)
* Add controller side of implementation

This new code implements the sensor to "detect" cars incoming to change the signal for them

* Improve timing

Due to an odd bug, the micro:bits won't actually take 1 second if I time the loop to do that.

Instead I've changed it to loop for 700 ms, which improves the lagging closer to 1 second. Haven't verified if its actually closer just feels closer haha.